Das U-Boot fixes
The RPi firmware uses a somewhat dynamic address to load the fdt and
passes this address along.  U-Boot, however, only populates the fdt_addr
environment variable with this address if fdt_addr is not already
defined.  The previous script was saving fdt_addr as part of the
environment, so would fail if the firmware load address ever changed.

Assorted fixes to deal with this issue:
* Update u-boot.ush to never save fdt_addr with the environment
* Generate a default u-boot environment file using Buildroot
* Write the default u-boot environment to sdcard.img (both env slots)

Other cleanup in u-boot.ush:
* Switch to using local variables where possible
* Remove temporary environment variables after use
* Cleanup inconsistent whitespace (convert spaces to tabs)

Signed-off-by: Charles Steinkuehler <cstein@newtek.com>