U-boot patch: Copy RPI fw modified fdt settings to u-boot loaded fdt
The closed source RPI firmware modifies several entries in the
device-tree that are important for proper boot. The complete list
of these entries can be found at:

https://forums.raspberrypi.com/viewtopic.php?p=1974233&hilit=firmware+device+tree+dma+ranges#p1974233

The function ft_board_setup() was modified so that if u-boot has
loaded a fdt that is different than the rpi fw fdt, then
it will copy all those important nodes and properties
created and modified by the rpi fw to the u-boot loaded fdt
blob.

Signed-off-by: Leonardo Amorim <lam@vizrt.com>